+++ RELEASENOTES.md
@@ -2,7 +2,8 @@
 
 ### dev-v2 (not yet released) ###
 
-* Support 32-bit PCM float output from `DefaultAudioSink`.
+* Support 32-bit PCM float output from `DefaultAudioSink`, and add an option to
+  use this with `FfmpegAudioRenderer`.
 
 ### 2.6.0 ###